Group 11 backend sale route

+++ back-end/src/database/models/Product.js
@@ -4,7 +4,7 @@ const ProductModel = (sequelize, DataTypes) => {
     {
       id: { autoIncrement: true, primaryKey: true, type: DataTypes.INTEGER },
       name: { type: DataTypes.STRING },
-      price: { type: DataTypes.DOUBLE },
+      price: { type: DataTypes.DECIMAL(4, 2) },
       urlImage: { type: DataTypes.STRING, field: 'url_image' },
     },
     { timestamps: false, tableName: 'products', underscored: true },
